Refactoring pod listing into a separate component
