+++ src/WebUI/Components/PodsComponent.scss
@@ -1,0 +1,5 @@
+@import "./Common.scss";
+
+.pl-details {
+    margin-top: 10px
+}
